Cache fix on upload

+++ src/app/pages/upload/upload.page.ts
@@ -1,33 +1,33 @@
-import { ActivatedRoute, Router } from '@angular/router';
-import { BaseDirectory, mkdir, readDir, remove, writeFile } from '@tauri-apps/plugin-fs';
-import { ChangeDetectorRef, Component, ElementRef, QueryList, ViewChild, ViewChildren, signal } from '@angular/core';
 import { CommonModule, Location } from '@angular/common';
+import { ChangeDetectorRef, Component, ElementRef, QueryList, ViewChild, ViewChildren, signal } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 import { IonButton, IonCard, IonCardContent, IonCardHeader, IonCardSubtitle, IonCardTitle, IonCol, IonContent, IonGrid, IonIcon, IonImg, IonInput, IonItem, IonLabel, IonList, IonRow, IonSpinner, ModalController } from '@ionic/angular/standalone';
+import { BaseDirectory, mkdir, readDir, remove, writeFile } from '@tauri-apps/plugin-fs';
+import { addOutline, checkmarkCircle, closeCircle, folder, logoSoundcloud, logoYoutube, removeOutline, trashOutline } from 'ionicons/icons';
 import { Measures, MusicDto, NoteDataDto } from 'src/app/game/game-model/music.dto';
 import { MusicOrigin, MusicPlayerCommon } from 'src/app/game/music-player/IMusicPlayer';
 import { SccReader, SccWriter } from './reader.ssc';
-import { addOutline, checkmarkCircle, closeCircle, folder, logoSoundcloud, logoYoutube, removeOutline, trashOutline } from 'ionicons/icons';
 
-import { AppComponent } from 'src/app/app.component';
-import { DanceType } from './../../game/constants/dance-type.enum';
 import { FormsModule } from '@angular/forms';
-import { HeaderComponent } from 'src/app/shared/component/header/header.component';
+import { addIcons } from 'ionicons';
+import { AppComponent } from 'src/app/app.component';
+import { MusicPlayerLocalComponent } from "src/app/game/music-player/music-player-local/music-player-local.component";
+import { MusicFirestoreService } from 'src/app/services/firestore/music.firestore.service';
+import { UserFirestoreService } from 'src/app/services/firestore/user.firestore.service';
+import { musicLocalService } from 'src/app/services/local-storage/local.music.service';
 import { MusicCacheService } from 'src/app/services/local-storage/music.cache.service';
+import { HeaderComponent } from 'src/app/shared/component/header/header.component';
+import { RadarScoreComponent } from "src/app/shared/component/radar-score/radar-score.component";
+import isTauri from 'src/app/shared/utils/tauri';
+import { MusicPlayerSoundcloudComponent } from "../../game/music-player/music-player-soundcloud/music-player-soundcloud.component";
+import { MusicPlayerYoutubeComponent } from "../../game/music-player/music-player-youtube/music-player-youtube.component";
+import { DanceType } from './../../game/constants/dance-type.enum';
+import { NoteDifficulty } from './../../game/constants/note-difficulty.enum';
 import { MusicEditableFieldAutocompleteComponent } from './editable-field-autocomplete/editable-field-autocomplete.component';
 import { MusicEditableFieldComponent } from './editable-field/editable-field.component';
 import { MusicEditableListComponent } from './editable-list/editable-list.component';
-import { MusicFirestoreService } from 'src/app/services/firestore/music.firestore.service';
-import { MusicPlayerLocalComponent } from "src/app/game/music-player/music-player-local/music-player-local.component";
-import { MusicPlayerSoundcloudComponent } from "../../game/music-player/music-player-soundcloud/music-player-soundcloud.component";
-import { MusicPlayerYoutubeComponent } from "../../game/music-player/music-player-youtube/music-player-youtube.component";
 import { MusicSelectComponent } from './select/select.component';
-import { NoteDifficulty } from './../../game/constants/note-difficulty.enum';
-import { RadarScoreComponent } from "src/app/shared/component/radar-score/radar-score.component";
 import { TestNoteComponent } from './test-note/test-note.component';
-import { UserFirestoreService } from 'src/app/services/firestore/user.firestore.service';
-import { addIcons } from 'ionicons';
-import isTauri from 'src/app/shared/utils/tauri';
-import { musicLocalService } from 'src/app/services/local-storage/local.music.service';
 
 @Component({
   selector: 'app-upload',
@@ -117,8 +117,6 @@ export class UploadPage {
     }
 
     this.notesDirty.set(dirtyArray);
-
-    console.log("Updated dirty")
   }
 
   anyNoteDirty(): boolean {
@@ -142,7 +140,7 @@ export class UploadPage {
   }
 
   async loadDbMusic(musicId: string) {
-    const data = await this.musicCacheService.getMusic(musicId);
+    const data = await this.musicCacheService.getMusicWithCheck(musicId);
     if (data) {
       this.musicDataDb = data;
       this.isEditDB.set(true);
@@ -303,21 +301,23 @@ export class UploadPage {
 
   //#region Upload
   async uploadMusic(): Promise<void> {
-    if (this.musicData === null) return;
-    if (this.isEditDB()) {
-      await this.fireStoreService.updateMusic(this.musicData)
-      AppComponent.presentOkToast("Music updated successfully!");
+    if (this.musicData === null || this.isEditDB()) return;
 
-      await this.loadDbMusic(this.musicData.id);
-      if (!this.anyNoteDirty()) {
-        this.deleteFile(true)
-      }
-    }
-    else {
-      await this.fireStoreService.uploadMusic(this.musicData)
-      this.isEditDB.set(true);
-      AppComponent.presentOkToast("Music uploaded successfully!");
-      this.switchToDbPage();
+    await this.musicCacheService.createMusicInBase(this.musicData)
+    this.isEditDB.set(true);
+    AppComponent.presentOkToast("Music uploaded successfully!");
+    this.switchToDbPage();
+  }
+
+  async updateMusic(): Promise<void> {
+    if (this.musicData === null || !this.isEditDB()) return;
+
+    await this.musicCacheService.updateMusicInBase(this.musicData)
+    AppComponent.presentOkToast("Music updated successfully!");
+
+    await this.loadDbMusic(this.musicData.id);
+    if (!this.anyNoteDirty()) {
+      this.deleteFile(true)
     }
   }
 
@@ -355,6 +355,7 @@ export class UploadPage {
       if (!silent) AppComponent.presentOkToast("Music has been discarded");
 
     } else {
+      this.musicCacheService.updateCache().then(() => this.loadDbMusic(musicId));
       if (!silent) AppComponent.presentOkToast("Music changes have been removed")
     }
 
@@ -362,7 +363,7 @@ export class UploadPage {
 
   async deleteMusic() {
     if (!this.musicDataDb) return;
-    await this.fireStoreService.deleteMusic(this.musicDataDb.id);
+    await this.musicCacheService.deleteMusicInBase(this.musicDataDb.id);
     this.musicDataDb = null;
     this.isEditDB.set(false);
     this.cd.detectChanges()
